Load or attach 'RAVE' subject

Load or attach 'RAVE' a subject to current analysis context. The goal is to allow analysis and visualization to be focused on one that subject.

rave_load(
  subject,
  epoch,
  reference,
  before_onset,
  after_onset,
  electrodes,
  attach = FALSE
)

rave_attach(x, ...)

# S3 method for RAVERepository
rave_attach(x, ...)

# S3 method for character
rave_attach(x, epoch, reference, before_onset, after_onset, electrodes, ...)

# S3 method for RAVESubject
rave_attach(x, epoch, reference, before_onset, after_onset, electrodes, ...)

attached_repo(test = FALSE)

rave_attached

Arguments

subject
character, the format is project/subject, for example "demo/DemoSubject"

epoch
character, name of epoch, located in subject's 'meta' folder, with file name "epoch_<name>.csv"

reference
character, name of reference scheme, located in subject's 'meta' folder, with file name "reference_<name>.csv"

before_onset
seconds before trial onset

after_onset
seconds after trial onset

electrodes
electrodes to load, if missing then default to all possible electrodes

attach
whether to attach loaded repository for analysis

x
either RAVERepository, RAVESubject instance or character, or NULL. For character, it's identical to subject; for NULL, current attached repository will be detached.

...
passed to other methods

test
whether return NULL instead of raising errors when no subject is attached

Format

An object of class ravebase_rave_attached (inherits from ravebase_readonly) of length 14.

Value

rave_load and rave_attach returns loaded repository, an instance of RAVERepository class. attached_repo returns currently loaded repository. If no repository loaded then it raises errors.

Details

rave_attached is a list providing various of methods that handles attached subjects. It also allows 'RAVE' modules designed for single-subject analysis to get information of what's to be analyzed.

epoch and reference are located in subject's meta folder. The folder path is "<RAVE data_path>/project/subject/rave/meta/". Epoch files start with "epoch_", and its name is decided by the characters following the prefix. For example, an epoch named "auditory" will have file name "epoch_auditory.csv". This rule applies to reference files, except that the prefix is "reference_"

before_onset and after_onset affects the amount of data to be loaded. For example, before_onset=1 and after_onset=2 will result in total 3 seconds of data to be loaded: 1 second before trial onset and 2 after onset. The reason to load data before onset is to leave certain time range for baseline. If the trial lasts for 2 seconds, loading 3 seconds of data allows around 1 second to be counted into baseline. The trial onset time for each trial is defined at column "Time" in the epoch file.

If attach=TRUE, then rave_load is equivalent to rave_attach.